requirements.txt
Brotli==1.1.0
certifi==2024.7.4
charset-normalizer==3.3.2
colorama==0.4.6
googleapis-common-protos==1.63.2
grpcio==1.65.5
idna==3.7
lz4==4.3.3
mutagen==1.47.0
pinecone-client==5.0.1
pinecone-plugin-inference==1.0.3
pinecone-plugin-interface==0.0.7
protobuf==4.25.4
protoc-gen-openapiv2==0.0.1
pycryptodomex==3.20.0
requests==2.32.3
tqdm==4.66.5
typing_extensions==4.12.2
Brotli==1.1.0
certifi==2024.7.4
charset-normalizer==3.3.2
colorama==0.4.6
fastapi==0.68.0
googleapis-common-protos==1.63.2
grpcio==1.65.5
urllib3==2.2.2
websockets==12.0
fastapi==0.68.0
uvicorn==0.15.0
pinecone-client==2.2.1
yt-dlp==2023.3.4
python-dotenv==0.19.0
sentence-transformers==2.2.2
pydantic==1.8.2
requests==2.26.0
tqdm==4.62.3
Brotli==1.1.0
certifi==2024.7.4
charset-normalizer==3.3.2
colorama==0.4.6
googleapis-common-protos==1.63.2
grpcio==1.65.5
idna==3.7
lz4==4.3.3
mutagen==1.47.0
pinecone-client==5.0.1
pinecone-plugin-inference==1.0.3
pinecone-plugin-interface==0.0.7
protobuf==4.25.4
protoc-gen-openapiv2==0.0.1
pycryptodomex==3.20.0
requests==2.32.3
tqdm==4.66.5
typing_extensions==4.12.2
Brotli==1.1.0
certifi==2024.7.4
charset-normalizer==3.3.2
colorama==0.4.6
fastapi==0.68.0
googleapis-common-protos==1.63.2
grpcio==1.65.5
urllib3==2.2.2
websockets==12.0
fastapi==0.68.0
uvicorn==0.15.0
pinecone-client==2.2.1
yt-dlp==2023.3.4
python-dotenv==0.19.0
sentence-transformers==2.2.2
pydantic==1.8.2
requests==2.26.0
tqdm==4.62.3

yt-dlp==2024.8.6
yt-dlp==2024.8.6

Brotli==1.1.0
certifi==2024.7.4
charset-normalizer==3.3.2
colorama==0.4.6
googleapis-common-protos==1.63.2
grpcio==1.65.5
idna==3.7
lz4==4.3.3
mutagen==1.47.0
pinecone-plugin-inference==1.0.3
pinecone-plugin-interface==0.0.7
protobuf==4.25.4
protoc-gen-openapiv2==0.0.1
pycryptodomex==3.20.0
typing_extensions==4.12.2
urllib3==2.2.2
websockets==12.0
fastapi==0.68.0
uvicorn==0.15.0
pinecone-client==2.2.1
yt-dlp==2023.3.4
python-dotenv==0.19.0
sentence-transformers==2.2.2
pydantic==1.8.2
requests==2.26.0
tqdm==4.62.3
Brotli==1.1.0
certifi==2024.7.4
charset-normalizer==3.3.2
colorama==0.4.6
googleapis-common-protos==1.63.2
grpcio==1.65.5
idna==3.7
lz4==4.3.3
mutagen==1.47.0
pinecone-plugin-inference==1.0.3
pinecone-plugin-interface==0.0.7
protobuf==4.25.4
protoc-gen-openapiv2==0.0.1
pycryptodomex==3.20.0
typing_extensions==4.12.2
urllib3==2.2.2
websockets==12.0
fastapi==0.68.0
uvicorn==0.15.0
pinecone-client==2.2.1
yt-dlp==2023.3.4
python-dotenv==0.19.0
sentence-transformers==2.2.2
pydantic==1.8.2
requests==2.26.0
tqdm==4.62.3

from fastapi import FastAPI, HTTPException, Query, BackgroundTasks
from fastapi.middleware.cors import CORSMiddleware
from pydantic import BaseModel, conlist, Field
import logging
import re
import os
from dotenv import load_dotenv